pyproject.toml
[project]
name = "fastapi-files"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.14",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]
